# server/kinematics_server/main.py
# generated by fastapi-codegen:
#   filename:  openapi.yaml
#   timestamp: 2025-09-12T15:49:16+00:00

from __future__ import annotations

from typing import List

from fastapi import FastAPI, Query
from fastapi.responses import RedirectResponse

from .models import WaboLinksGetResponseItem, WaboKinematikResults
from .handlers import get_cartesian_position_impl, get_link_information_impl, get_version_impl, get_kinematicresults_impl



app = FastAPI(
    title='WaBo API',
    license={
        'name': 'Apache 2.0',
        'url': 'https://www.apache.org/licenses/LICENSE-2.0.html',
    },
    description='Interact with robots in an easy and intuitive way.\n',
    contact={
        'name': 'Copyright (c) 2025 Wandelbots GmbH',
        'url': 'https://www.wandelbots.com',
    },
    version='0.1.0 beta',
    servers=[{'url': '/'}],
)

@app.get("/", include_in_schema=False)
def root():
    return RedirectResponse(url="/docs")

@app.get('/wabo/cartesian', response_model=List[float], tags=['WaBo'])
def get_cartesian_position(joint_position: List[float] = Query(...)) -> List[float]:
    return get_cartesian_position_impl(joint_position)

@app.get('/wabo/links', response_model=List[WaboLinksGetResponseItem], tags=['WaBo'])
def get_link_information() -> List[WaboLinksGetResponseItem]:
    return get_link_information_impl()

@app.get('/wabo/version', response_model=str, tags=['WaBo'])
def get_version() -> str:
    return get_version_impl()

@app.get('/wabo/getkinematicresults', response_model=WaboKinematikResults, tags=['WaBo'])
def get_kinematic_results(joint_position: List[float] = Query(...)) -> WaboKinematikResults:
    return get_kinematicresults_impl(joint_position)